split datasets.py, split LoadImagesAndLabels init,  add albumentations to requirements, train tested on coco -> keep working
