+++ requirements.txt
@@ -11,6 +11,7 @@ scipy>=1.4.1
 torch>=1.7.0
 torchvision>=0.8.1
 tqdm>=4.41.0
+albumentations
 
 # Logging -------------------------------------
 tensorboard>=2.4.1
