+++ utils/datasets/InfiniteDataLoader.py
@@ -1,0 +1,24 @@
+from torch.utils.data.dataloader import DataLoader
+# #####################################
+from .RepeatSampler import RepeatSampler
+# #####################################
+
+
+class InfiniteDataLoader(DataLoader):
+    """ Dataloader that reuses workers
+
+    Uses same syntax as vanilla DataLoader
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        object.__setattr__(self, 'batch_sampler', 
+            RepeatSampler(self.batch_sampler))
+        self.iterator = super().__iter__()
+
+    def __len__(self):
+        return len(self.batch_sampler.sampler)
+
+    def __iter__(self):
+        for i in range(len(self)):
+            yield next(self.iterator)
